+++ scheme/ComputeViscousFluxesAndStoreFunctor.cpp
@@ -63,7 +63,7 @@ ComputeViscousFluxesAndStoreFunctor<dim, device_t>::apply(ConfigMap const &     
   // Important note: the caller is responsible for provide a flux array with right shape.
   {
     [[maybe_unused]] auto flux_block_sizes = q_ghosted.block_size();
-    flux_block_sizes[direction]++;
+    flux_block_sizes[static_cast<size_t>(direction)]++;
     assertm(flux_block_sizes == fluxes.shape(), "Flux array has incompatible shape.");
   }
